Adiciona funcao de excluir conta na aba de Profile do tutor
